+++ packages/dara-core/dara/core/main.py
@@ -29,6 +29,7 @@ from fastapi import FastAPI, HTTPException, Request
 from fastapi.responses import HTMLResponse
 from fastapi.staticfiles import StaticFiles
 from prometheus_client import start_http_server
+from pydantic.json import ENCODERS_BY_TYPE
 from starlette.templating import Jinja2Templates, _TemplateResponse
 
 from dara.core.auth import auth_router
@@ -41,6 +42,7 @@ from dara.core.defaults import (
 )
 from dara.core.internal.cgroup import get_cpu_count, set_memory_limit
 from dara.core.internal.devtools import send_error_for_session
+from dara.core.internal.encoder_registry import encoder_registry
 from dara.core.internal.pool import TaskPool
 from dara.core.internal.registries import (
     action_def_registry,
@@ -226,6 +228,13 @@ def _start_application(config: Configuration):
     # Add WS handlers
     for kind, handler in config.ws_handlers.items():
         custom_ws_handlers_registry.register(kind, handler)
+
+    # update encoder registry
+    encoder_registry.update(config.encoders)
+
+    # Inject the encoder to pydantic ENCODERS_BY_TYPE, which will be called in fastapi jsonable_encoder
+    for key, value in encoder_registry.items():
+        ENCODERS_BY_TYPE[key] = value['serialize']
 
     # Generate a new build_cache
     try: