+++ packages/dara-core/changelog.md
@@ -2,6 +2,10 @@
 title: Changelog
 ---
 
+## NEXT
+
+-   Added serialize/deserialize support to numpy/pandas datatype
+-   Added custom encoder support. User can add encoder to handle serialization/deserialization of a type
 ## 1.1.6
 
 -   Internal: dedupe custom registry lookup calls